+++ Model/WallBlock.java
@@ -37,8 +37,8 @@ public class WallBlock extends GameObject {
 	public void paint(Graphics g, int BLOC_SIZE) {
 		g.setColor(getColor());
         g.fillRect(
-        		getPos().getX() * BLOC_SIZE-1,
-        		getPos().getY() * BLOC_SIZE-1,
+        		(int)(getPos().getX() * BLOC_SIZE-1),
+        		(int)(getPos().getY() * BLOC_SIZE-1),
         		(BLOC_SIZE * getSize().getWidth()),
         		(BLOC_SIZE * getSize().getHeight()));
 	}
